"Steak Holder" (#3)
Okay, so I don't actually know how HTML or Github works, so I don't know if this will actually do things or not, but this typo is on the website and it really bothers me. Unless, of course, team 2056 is the best in FIRST at holding pieces of meat for great lengths of time.